==> scripts/07.1_linear-models_subplot-Count.R <==
# Created: 2024-05-26
# Last updated: 2024-08-21

# Purpose: Run generalized linear models for subplot data, with Count as response variable. 
#   Check for overdispersion and zero-inflation.

# Presence of overdispersion and for Count indicates a negative binomial should be used.
# Including (1 | Site / Plot) accounts for the non-independence of repeat sampling.

# "All variables" includes: Perc_dev_cum, AridityIndex, Treatment, PlantSource2,
#     PlotMix_Climate, Duration, Lifeform, MAT, MAP, Sand_content, Cum_precip
# However, AridityIndex and MAP are usually too correlated to include both, so MAP often dropped.


library(tidyverse)
library(glmmTMB)
library(performance)
library(DHARMa)

# Load data ---------------------------------------------------------------

subplot.raw <- read_csv("data/cleaned/04.1_subplot-data_clean.csv")
prism.data <- read_csv("data/cleaned/03.2_monitoring-events-with-PRISM-climate-data_clean.csv")
cum.pd <- read_csv("data/cleaned/03.3_cumulative-precip_percent-deviation-from-norm_clean.csv")
ai <- read_csv("data/cleaned/03.4_aridity-index-values_clean.csv")
siteplot.id <- read_csv("data/cleaned/02_SitePlotID_clean.csv")


# Data wrangling ----------------------------------------------------------

## Combine ----------------------------------------------------------------

# Reorganize columns for left_join()
cum.pd.subplot <- cum.pd |> 
  select(Region, Site, SiteDateID, Date_Seeded, Date_Monitored, Perc_deviation, Deviation_mm) |> 
  rename(Perc_dev_cum = Perc_deviation,
         Dev_mm_cum = Deviation_mm)

# Combine all variables
subplot <- subplot.raw |> 
  left_join(prism.data) |> 
  left_join(ai) |> 
  left_join(cum.pd.subplot)

# Check for NAs
apply(subplot, 2, anyNA)


## Remove Inf from Perc_dev_cum -------------------------------------------

# Infinity created when there was no rain during the monitoring period. This only happens
#   twice and these instances can be dropped.
subplot <- subplot |> 
  filter(Perc_dev_cum != Inf)



## Re-level categorical variables to set reference ------------------------

# Treatment
unique(subplot$Treatment)
subplot$Treatment <- as.factor(subplot$Treatment)
subplot$Treatment <- relevel(subplot$Treatment, ref = "Control")

# PlotMix_Climate
unique(subplot$PlotMix_Climate)
#   Allow Current to remain reference; because of uneven sample sizes, if
#     None is reference then Projected will be dropped from models. Better
#     to drop None and be able to compare Current & Projected.

# Create subplot data with 0 as ref for PlantSource2, Weedy, Duration, Lifeform
#   Change reference for Treatment
subplot0 <- subplot


# Change ref from 0 for better comparisons
#   Lifeform
unique(subplot$Lifeform)
subplot$Lifeform <- as.factor(subplot$Lifeform)
subplot$Lifeform <- relevel(subplot$Lifeform, ref = "Forb")
#   Relevel to see if Forb is different than Grass (vs. different than 0)

# Duration
unique(subplot$Duration)
subplot$Duration <- as.factor(subplot$Duration)
subplot$Duration <- relevel(subplot$Duration, ref = "Annual")
#   Relevel to see if Annual is different than Perennial (vs. different than 0)

# Weedy
unique(subplot$Weedy)
subplot$Weedy <- as.factor(subplot$Weedy)
subplot$Weedy <- relevel(subplot$Weedy, ref = "Weedy")




## Remove Pellets ---------------------------------------------------------

# Pellets were only used at Mojave
subplot |> 
  filter(Treatment == "Pellets") |> 
  select(Region, Site) |> 
  distinct(.keep_all = TRUE)

# All
nopellet <- subplot |> 
  filter(Treatment != "Pellets")
nopellet0 <- subplot0 |> 
  filter(Treatment != "Pellets")

# Desirable
nopellet.des <- nopellet |> 
  filter(Weedy != "Weedy")

# Weedy
nopellet.weed <- nopellet |> 
  filter(Weedy != "Desirable")


## Separate out Pits, Seed and Control only -------------------------------

# All
pitseed <- subplot |> 
  filter(Treatment %in% c("Control", "Seed", "Pits"))
pitseed0 <- subplot0 |> 
  filter(Treatment %in% c("Control", "Seed", "Pits"))

# Desirable
pitseed.des <- pitseed |> 
  filter(Weedy != "Weedy")

# Weedy
pitseed.weed <- pitseed |> 
  filter(Weedy != "Desirable")


## Separate out Sonoran sites (6) -----------------------------------------

# All
sonoran <- subplot |> 
  filter(Region %in% c("Sonoran SE", "Sonoran Central"))

# Desirable
sonoran.des <- sonoran |> 
  filter(Weedy != "Weedy")

# Weedy
sonoran.weed <- sonoran |> 
  filter(Weedy != "Desirable")


## Separate out Northern AZ sites (8) -------------------------------------

# All
naz <- subplot |> 
  filter(Region == "Colorado Plateau")

# Desirable
naz.des <- naz |> 
  filter(Weedy != "Weedy")

# Weedy
naz.weed <- naz |> 
  filter(Weedy != "Desirable")


## Separate out Utah sites (3) --------------------------------------------

# All
utah <- subplot |> 
  filter(Region == "Utah")

# Desirable
utah.des <- utah |> 
  filter(Weedy != "Weedy")

# Weedy
utah.weed <- utah |> 
  filter(Weedy != "Desirable")




# pitseed dataset (Control, Seed, Pits) -----------------------------------

# Control, Seed, and Pits plots included (Pellets, Mulch & ConMod excluded)

## Poisson ----------------------------------------------------------------

# All variables, no random effects, 0 as ref
pos.pitseed0.0 <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                     PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                   data = pitseed0, family = "poisson")
summary(pos.pitseed0.0)
check_overdispersion(pos.pitseed0.0) # overdispersion detected
check_zeroinflation(pos.pitseed0.0) # no zero-inflation detected
check_model(pos.pitseed0.0)

# All variables, ref adjusted, no random effects: does not converge
pos.pitseed.0 <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                       PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                     data = pitseed, family = "poisson") # did not converge

# All variables, 0 as ref, with random effects: does not converge
pos.pitseed0 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                        PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                        Cum_precip + (1 | Site / Plot),
                      data = pitseed0,
                      family = genpois) # did not converge



## Quasi-Poisson ----------------------------------------------------------

# All variables, 0 as ref, no random effects
qpos.pitseed0.0 <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                       PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                     data = pitseed0, family = "quasipoisson")
summary(qpos.pitseed0.0)
check_model(qpos.pitseed0.0)
check_overdispersion(qpos.pitseed0.0) # overdispersion detected
check_zeroinflation(qpos.pitseed0.0) # no zero-inflation detected

# All variables, ref adjusted, no random effects: did not converge
qpos.pitseed.0 <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                         PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                       data = pitseed, family = "quasipoisson") # does not converge


## Negative binomial ------------------------------------------------------

# All variables, 0 as ref, no random effects
nb.pitseed0.0 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                        PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                        Cum_precip,
                      data = pitseed0,
                      family = nbinom2)
summary(nb.pitseed0.0)
res.nb.pitseed0.0 <- simulateResiduals(nb.pitseed0.0)
check_model(nb.pitseed0.0)
check_zeroinflation(nb.pitseed0.0) # no zero-inflation
testZeroInflation(res.nb.pitseed0.0) # no zero-inflation
check_overdispersion(nb.pitseed0.0) # overdispersion detected
plotResiduals(res.nb.pitseed0.0)
plotQQunif(res.nb.pitseed0.0)
outliers(res.nb.pitseed0.0)


# All variables, 0 as ref, nested random effect of Site/Plot
nb.pitseed0 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                        PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                        Cum_precip + (1 | Site / Plot),
                      data = pitseed0,
                      family = nbinom2)
summary(nb.pitseed0)
res.nb.pitseed0 <- simulateResiduals(nb.pitseed0)
check_model(nb.pitseed0)
check_overdispersion(nb.pitseed0) # indicates overdispersion still
check_zeroinflation(nb.pitseed0) # no zero-inflation
step(nb.pitseed0)
testDispersion(res.nb.pitseed0, alternative = "greater") # indicates overdispersion still
plotResiduals(nb.pitseed0)

# All variables, ref adjusted, nested random effect of Site/Plot
nb.pitseed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                        PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                        Cum_precip + (1 | Site / Plot),
                      data = pitseed,
                      family = nbinom2)
summary(nb.pitseed)
res.nb.pitseed <- simulateResiduals(nb.pitseed)
check_model(nb.pitseed)
check_overdispersion(nb.pitseed) # indicates overdispersion still
step(nb.pitseed)
testDispersion(res.nb.pitseed, alternative = "greater") # indicates overdispersion still
plotResiduals(nb.pitseed)


# 1: Drop MAP (collinearity) and change PlantSource2 to Weedy
nb.pitseed1 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + Weedy + 
                        PlotMix_Climate + Duration + Lifeform + MAT + Sand_content + 
                        Cum_precip + (1 | Site / Plot),
                      data = pitseed,
                      family = nbinom2)
summary(nb.pitseed1)
res.nb.pitseed1 <- simulateResiduals(nb.pitseed1)
check_model(nb.pitseed1)
plotResiduals(nb.pitseed1)
plotQQunif(nb.pitseed1)


## Zero-inflated negative binomial ----------------------------------------

# All explanatory variables, ref as 0, no random effects: does not converge
zinb.pitseed0.0 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                          PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                        data = pitseed0,
                        family = nbinom2,
                        ziformula = ~.) # did not converge



# All data ----------------------------------------------------------------

## Poisson ----------------------------------------------------------------

# All variables, ref adjusted, no random effects
pos.all.0 <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                     PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                   data = subplot, family = "poisson")
summary(pos.all.0)
check_overdispersion(pos.all.0) # overdispersion detected
check_zeroinflation(pos.all.0) # no zero-inflation detected

# All variables, ref adjusted, with random effects: does not converge
pos.all <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                         PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                         Cum_precip + (1 | Site / Plot),
                       data = subplot,
                       family = genpois) # did not converge


## Negative binomial ------------------------------------------------------

# All variables, ref adjusted, no random effects
#   Perc_cum_dev significant
nb.all.0 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                          PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                          Cum_precip,
                        data = subplot,
                        family = nbinom2)
summary(nb.all.0)
check_overdispersion(nb.all.0) # overdispersion detected
check_zeroinflation(nb.all.0) # no zero-inflation detected


# All variables, ref adjusted, nested random effect of Site/Plot
#   Perc_cum_dev not significant
nb.all <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                      PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                      Cum_precip + (1 | Site / Plot),
                    data = subplot,
                    family = nbinom2)
summary(nb.all)
res.nb.all <- simulateResiduals(nb.all)
check_overdispersion(nb.all) # overdispersion detected
check_zeroinflation(nb.all) # no zero-inflation detected
check_collinearity(nb.all)



# nopellet dataset --------------------------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, ref adjusted, nested Site/Plot as random effect
nb.nopellet <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                              PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                              Cum_precip + (1 | Site / Plot),
                            data = nopellet,
                            family = nbinom2)
summary(nb.nopellet)
res.nb.nopellet <- simulateResiduals(nb.nopellet)
check_overdispersion(nb.nopellet) # overdispersion detected
check_zeroinflation(nb.nopellet) # no zero-inflation detected
check_collinearity(nb.nopellet)
plotQQunif(res.nb.nopellet)
testOutliers(res.nb.nopellet)
outliers(res.nb.nopellet)
plotResiduals(res.nb.nopellet)


# All variables, ref adjusted, SitePlotID as random effect
nb.nopellet.1 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                         PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                         Cum_precip + (1 | SitePlotID),
                       data = nopellet,
                       family = nbinom2)
summary(nb.nopellet.1)


# All variables, ref adjusted, no random effect: produces NaNs
nb.nopellet0 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                         PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                         Cum_precip,
                       data = nopellet,
                       family = nbinom2)






# pitseed subset by Weedy/Desirable ---------------------------------------

## Poisson ----------------------------------------------------------------

# All variables, no random effects: Desirable
pos.pitseed00.des <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                           PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                         data = pitseed.des, family = "poisson")
summary(pos.pitseed00.des)
res.pos.pitseed00.des <- simulateResiduals(pos.pitseed00.des)
check_overdispersion(pos.pitseed00.des) # overdispersion detected
check_zeroinflation(pos.pitseed00.des) # no zero-inflation detected
plotQQunif(res.pos.pitseed00.des)
plotResiduals(res.pos.pitseed00.des)

# All variables, no random effects: Weedy
pos.pitseed00.weed <- glm(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 +
                           PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + Cum_precip,
                         data = pitseed.weed, family = "poisson")
summary(pos.pitseed00.weed)
res.pos.pitseed00.weed <- simulateResiduals(pos.pitseed00.weed)
check_overdispersion(pos.pitseed00.weed) # overdispersion detected
check_zeroinflation(pos.pitseed00.weed) # no zero-inflation detected


## Negative binomial ------------------------------------------------------

# All variables, no random effects: Desirable
nb.pitseed00.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                          PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                          Cum_precip,
                        data = pitseed.des,
                        family = nbinom2)
summary(nb.pitseed00.des)
res.nb.pitseed00.des <- simulateResiduals(nb.pitseed00.des)
check_overdispersion(nb.pitseed00.des) # overdispersion detected
check_zeroinflation(nb.pitseed00.des) # no zero-inflation detected
plotQQunif(res.nb.pitseed00.des)
testOutliers(res.nb.pitseed00.des)
outliers(res.nb.pitseed00.des)
plotResiduals(res.nb.pitseed00.des)

# All variables, no random effects: Weedy
nb.pitseed00.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                              PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                              Cum_precip,
                            data = pitseed.weed,
                            family = nbinom2)
summary(nb.pitseed00.weed)
res.nb.pitseed00.weed <- simulateResiduals(nb.pitseed00.weed)
check_overdispersion(nb.pitseed00.weed) # overdispersion detected
check_zeroinflation(nb.pitseed00.weed) # no zero-inflation detected
plotQQunif(res.nb.pitseed00.weed)
testOutliers(res.nb.pitseed00.weed)
outliers(res.nb.pitseed00.weed)
plotResiduals(res.nb.pitseed00.weed)



# nopellet subset by Weedy/Desirable --------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, ref adjusted, nested random effect of Site/Plot: Desirable
nb.nopellet.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                             PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                             Cum_precip + (1 | Site / Plot),
                           data = nopellet.des,
                           family = nbinom2)
summary(nb.nopellet.des)
res.nb.nopellet.des <- simulateResiduals(nb.nopellet.des)
plotQQunif(res.nb.nopellet.des)
check_overdispersion(nb.nopellet.des) # overdispersion detected
step(nb.nopellet.des)

# All variables, ref adjusted, nested random effect of Site/Plot: Weedy
nb.nopellet.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                             PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                             Cum_precip + (1 | Site / Plot),
                           data = nopellet.weed,
                           family = nbinom2)
summary(nb.nopellet.weed)
res.nb.nopellet.weed <- simulateResiduals(nb.nopellet.weed)
plotQQunif(res.nb.nopellet.weed)
plotResiduals(res.nb.nopellet.weed)
check_overdispersion(nb.nopellet.weed) # overdispersion detected


# 1: drop PlantSource2 and PlotMix_Climate: Desirable
nb.nopellet01.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment +  
                             Duration + Lifeform + MAT + MAP + Sand_content + 
                             Cum_precip + (1 | Site / Plot),
                           data = nopellet.des,
                           family = nbinom2)
summary(nb.nopellet01.des)
res.nb.nopellet01.des <- simulateResiduals(nb.nopellet01.des)
plotQQunif(res.nb.nopellet01.des)
plotResiduals(res.nb.nopellet01.des)
check_model(nb.nopellet01.des)
step(nb.nopellet01.des)

# 1: drop PlantSource2 and PlotMix_Climate: Weedy
nb.nopellet01.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment +  
                               Duration + Lifeform + MAT + MAP + Sand_content + 
                               Cum_precip + (1 | Site / Plot),
                             data = nopellet.weed,
                             family = nbinom2)
summary(nb.nopellet01.weed)
res.nb.nopellet01.weed <- simulateResiduals(nb.nopellet01.weed)
plotQQunif(res.nb.nopellet01.weed)
plotResiduals(res.nb.nopellet01.weed)
check_model(nb.nopellet01.weed)
step(nb.nopellet01.weed)


# 2: drop PlotMix_Climate, replace PlantSource2 with SpeciesSeeded: Desirable
nb.nopellet02.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + SpeciesSeeded +  
                               Duration + Lifeform + MAT + MAP + Sand_content + 
                               Cum_precip + (1 | Site / Plot),
                             data = nopellet.des,
                             family = nbinom2)
summary(nb.nopellet02.des)
res.nb.nopellet02.des <- simulateResiduals(nb.nopellet02.des)
plotQQunif(res.nb.nopellet02.des)
plotResiduals(res.nb.nopellet02.des)
check_model(nb.nopellet02.des)
step(nb.nopellet02.des)

# 2: drop PlotMix_Climate: Weedy
nb.nopellet02.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                                Duration + Lifeform + MAT + MAP + Sand_content + 
                                Cum_precip + (1 | Site / Plot),
                              data = nopellet.weed,
                              family = nbinom2)
summary(nb.nopellet02.weed)
res.nb.nopellet02.weed <- simulateResiduals(nb.nopellet02.weed)
plotQQunif(res.nb.nopellet02.weed)
plotResiduals(res.nb.nopellet02.weed)
check_model(nb.nopellet02.weed)
step(nb.nopellet02.weed)



# Sonoran sites -----------------------------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, nested random effect of Site/Plot: produces NaNs
nb.sonoran <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                            PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                            Cum_precip + (1 | Site / Plot),
                          data = sonoran,
                          family = nbinom2) # produces NaNs

# 1: Drop MAT & MAP (for collinearity): Desirable
nb.sonoran01 <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                              PlotMix_Climate + Duration + Lifeform + Sand_content + 
                              Cum_precip + (1 | Site / Plot),
                            data = sonoran,
                            family = nbinom2) 
summary(nb.sonoran01)
res.nb.sonoran01 <- simulateResiduals(nb.sonoran01)
plotQQunif(res.nb.sonoran01)
plotResiduals(res.nb.sonoran01)
check_overdispersion(nb.sonoran01) # overdispersion detected
check_zeroinflation(nb.sonoran01) # no zero-inflation detected
check_model(nb.sonoran01)
step(nb.sonoran01)



# Sonoran sites by Weedy/Desirable ----------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, nested random effect of Site/Plot: Desirable
nb.sonoran.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                             PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                             Cum_precip + (1 | Site / Plot),
                           data = sonoran.des,
                           family = nbinom2)
summary(nb.sonoran.des)
res.nb.sonoran.des <- simulateResiduals(nb.sonoran.des)
plotQQunif(res.nb.sonoran.des)
plotResiduals(res.nb.sonoran.des)
check_model(nb.sonoran.des)
check_overdispersion(nb.sonoran.des) # overdispersion detected
check_zeroinflation(nb.sonoran.des) # no overdispersion detected
check_collinearity(nb.sonoran.des)
step(nb.sonoran.des)

# All variables, nested random effect of Site/Plot: Weedy
nb.sonoran.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                            PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                            Cum_precip + (1 | Site / Plot),
                          data = sonoran.weed,
                          family = nbinom2)
summary(nb.sonoran.weed)
res.nb.sonoran.weed <- simulateResiduals(nb.sonoran.weed)
plotQQunif(res.nb.sonoran.weed)
plotResiduals(res.nb.sonoran.weed)
check_overdispersion(nb.sonoran.weed) # overdispersion detected
check_model(nb.sonoran.weed)
check_collinearity(nb.sonoran.weed)
step(nb.sonoran.weed)


# 1: Drop MAP (for collinearity): Desirable
nb.sonoran1.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                            PlotMix_Climate + Duration + Lifeform + MAT + Sand_content + 
                            Cum_precip + (1 | Site / Plot),
                          data = sonoran.des,
                          family = nbinom2)
summary(nb.sonoran1.des)
res.nb.sonoran1.des <- simulateResiduals(nb.sonoran1.des)
plotQQunif(res.nb.sonoran1.des)
plotResiduals(res.nb.sonoran1.des)
check_model(nb.sonoran1.des)
check_overdispersion(nb.sonoran1.des) # overdispersion detected
check_zeroinflation(nb.sonoran1.des) # no zero-inflation detected
check_collinearity(nb.sonoran1.des)
step(nb.sonoran1.des)

# 1: Drop MAP & Duration (for collinearity): Weedy
nb.sonoran1.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                              PlotMix_Climate + Lifeform + MAT + Sand_content + 
                              Cum_precip + (1 | Site / Plot),
                            data = sonoran.weed,
                            family = nbinom2)
summary(nb.sonoran1.weed)
res.nb.sonoran1.weed <- simulateResiduals(nb.sonoran1.weed)
plotQQunif(res.nb.sonoran1.weed)
plotResiduals(res.nb.sonoran1.weed)
check_model(nb.sonoran1.weed)
check_overdispersion(nb.sonoran1.weed) # overdispersion detected
check_zeroinflation(nb.sonoran1.weed) # no zero-inflation detected
check_collinearity(nb.sonoran1.weed)
step(nb.sonoran1.weed)



# Northern AZ sites -------------------------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, nested random effect of Site/Plot
nb.naz <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                        PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                        Cum_precip + (1 | Site / Plot),
                      data = naz,
                      family = nbinom2) 
summary(nb.naz)
res.nb.naz <- simulateResiduals(nb.naz)
plotQQunif(res.nb.naz)
plotResiduals(res.nb.naz)
check_model(nb.naz)
check_overdispersion(res.nb.naz)
check_collinearity(nb.naz)


# Northern AZ sites by Weedy/Desirable ------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, nested random effect of Site/Plot: Desirable
nb.naz.des <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                        PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                        Cum_precip + (1 | Site / Plot),
                      data = naz.des,
                      family = nbinom2)
summary(nb.naz.des)
res.nb.naz.des <- simulateResiduals(nb.naz.des)
plotQQunif(res.nb.naz.des)
plotResiduals(res.nb.naz.des)
check_model(nb.naz.des)
check_overdispersion(nb.naz.des) # overdispersion detected
check_zeroinflation(nb.naz.des) # no overdispersion detected
check_collinearity(nb.naz.des)
step(nb.naz.des)

# All variables, nested random effect of Site/Plot: Weedy
nb.naz.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                         PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                         Cum_precip + (1 | Site / Plot),
                       data = naz.weed,
                       family = nbinom2)
summary(nb.naz.weed)
res.nb.naz.weed <- simulateResiduals(nb.naz.weed)
plotQQunif(res.nb.naz.weed)
plotResiduals(res.nb.naz.weed)
check_model(nb.naz.weed)
check_overdispersion(nb.naz.weed) # overdispersion detected
check_collinearity(nb.naz.weed)
step(nb.naz.weed)

# 1: Drop PlantSource (for collinearity): Weedy
nb.naz1.weed <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment +  
                         PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                         Cum_precip + (1 | Site / Plot),
                       data = naz.weed,
                       family = nbinom2)
summary(nb.naz1.weed)
res.nb.naz1.weed <- simulateResiduals(nb.naz1.weed)
plotQQunif(res.nb.naz1.weed)
plotResiduals(res.nb.naz1.weed)
check_model(nb.naz1.weed)
check_overdispersion(nb.naz1.weed) # overdispersion detected
check_collinearity(nb.naz1.weed)


# 2: Perc_dev_cum as cubic: Desirable
nb.naz2.des <- glmmTMB(Count ~ I(Perc_dev_cum ^ 3) + AridityIndex + Treatment + PlantSource2 + 
                          PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                          Cum_precip + (1 | Site / Plot),
                        data = naz.des,
                        family = nbinom2)
summary(nb.naz2.des)
res.nb.naz2.des <- simulateResiduals(nb.naz2.des)
plotQQunif(res.nb.naz2.des)
plotResiduals(res.nb.naz2.des)
check_overdispersion(nb.naz2.des) # overdispersion detected
check_zeroinflation(nb.naz2.des) # no zero-inflation detected
check_model(nb.naz2.des)
step(nb.naz2.des)

# 2: Perc_dev_cum as cubic, drop PlantSource (for collinearity): Weedy
nb.naz2.weed <- glmmTMB(Count ~ I(Perc_dev_cum ^ 3) + AridityIndex + Treatment +  
                           PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                           Cum_precip + (1 | Site / Plot),
                         data = naz.weed,
                         family = nbinom2)
summary(nb.naz2.weed)
res.nb.naz2.weed <- simulateResiduals(nb.naz2.weed)
plotQQunif(res.nb.naz2.weed)
plotResiduals(res.nb.naz2.weed)
check_overdispersion(nb.naz2.weed) # overdispersion detected
check_zeroinflation(nb.naz2.weed) # no zero-inflation detected
check_model(nb.naz2.weed)
step(nb.naz2.weed)



# Utah sites --------------------------------------------------------------

## Negative binomial ------------------------------------------------------

# All variables, nested random effect of Site/Plot
nb.utah <- glmmTMB(Count ~ Perc_dev_cum + AridityIndex + Treatment + PlantSource2 + 
                    PlotMix_Climate + Duration + Lifeform + MAT + MAP + Sand_content + 
                    Cum_precip + (1 | Site / Plot),
                  data = utah,
                  family = nbinom2) # did not converge


save.image("RData/07.1_linear-models_subplot-Count.RData")
